ref(rtcp): Makes the CNAMERegistry an inner class.
